+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Also check this project's [releases](https://github.com/powerhome/redis-operator
 
 ### Changed
 - [Use the new docker bake tooling to build the developer tools image and remove vestigial development targets from the Makefile](https://github.com/powerhome/redis-operator/pull/31).
+- HAProxy resources are now named using the pattern `rfrh-[redis-failover-name]`
+- The Redis headless service is now named using the pattern `rfrh-[redis-failover-name]`
 
 ## [v1.8.0-rc2] - 2023-12-20